Fix test for receive default command (#995)
The test shall test that the ZigBeeNetworkManager does not notify about
default responses. However, after ignoring endpoints other than 1 the
test does not test this anymore, because it uses a default response to
endpoint 2 which is now dropped not because it is a default response,
but because it is not for endpoint 1.

This commit adapts the test to use a default response for endpoint 1, so
that it actually tests the correct behavior for default responses again.